Refactor populatedatabase function
