+++ backend/StreamServerApp/management/commands/populatedb.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from StreamServerApp.utils import delete_DB_Infos, populate_db_from_local_folder
+from StreamServerApp.database_utils import delete_DB_Infos, populate_db_from_local_folder
 from django.conf import settings
 
 class Command(BaseCommand):
